fix: build error 해결
Suspense 사용을 위한 page분리, middleware파일이 비어있어서 exoprt 추가
