+++ middleware.ts
@@ -1,0 +1,6 @@
+// src/middleware.ts
+import { NextRequest, NextResponse } from 'next/server'
+
+export function middleware(request: NextRequest) {
+  return NextResponse.next()
+}
