+++ src/features/challenge/components/challenge/group/create/DetailStep.tsx
@@ -6,7 +6,7 @@ import { useState } from 'react'
 import { UseFormReturn } from 'react-hook-form'
 import styled from '@emotion/styled'
 
-import { FullFormValues } from '@app/(with-header)/challenge/group/create/page'
+import { FullFormValues } from '@app/(with-header)/challenge/group/create/ChallengeCreatepage'
 import ErrorText from '@shared/components/errortext'
 import ImageInput from '@shared/components/image-input'
 import Loading from '@shared/components/loading'
